% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_label.R
\name{add_label}
\alias{add_label}
\title{Add label attribute to a vector.}
\usage{
add_label(.x, .label)
}
\arguments{
\item{.x}{A vector}

\item{.label}{A character vector of length 1. Label of the vector.}
}
\value{
A labelled vector (i.e. a vector with \code{label} attribute)
}
\description{
Add label attribute to a vector.
}
\examples{
\dontrun{
v <- 1:3
v <- add_label(v, "Some numbers")
v
attributes(v)
}
}
