% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Create project folders}
\usage{
create_folders(project_path = ".")
}
\arguments{
\item{project_path}{character}
}
\value{

}
\description{
Create project folders with a minimal structure
}
\examples{
temp_folder <- tempdir()
create_folders(temp_folder)
list.dirs(temp_folder)
}
