% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data.R
\name{describe_data}
\alias{describe_data}
\title{Describe data using {arsenal} package.}
\usage{
describe_data(
  dat,
  by = NULL,
  p_value = FALSE,
  total = FALSE,
  digits = 1L,
  digits.pct = 1L,
  digits.pvalue = 4L,
  ...
)
}
\arguments{
\item{dat}{A dataframe}

\item{by}{Grouping variable by which data are described. If NULL (default), overall total will be shown.}

\item{p_value}{A logical. Toggle to show p value (default is FALSE).}

\item{total}{A logical. Toggle to show overall total (default is FALSE) when grouping variable is defined.}

\item{digits}{An integer. Control the number of decimal place for continuous values (default is 1).}

\item{digits.pct}{An integer. Control the number of decimal place for percents (default is 1).}

\item{digits.pvalue}{An integer. Control the number of decimal place for p-values (default is 1).}

\item{...}{Arguments passed to tableby(). ?tableby for more details.}
}
\value{

}
\description{
This function creates a tableby object using more intuitive default
parameters. The tableby object should be passed to summary() to create a
descriptive table.
}
\examples{
\dontrun{
library(arsenal)
library(dplyr)

data(mockstudy)

mockstudy \%>\%
  select(arm, sex, age) \%>\%
  describe_data(by = arm) \%>\%
  summary(text = TRUE)

mockstudy \%>\%
  select(arm, sex, age) \%>\%
  describe_data() \%>\%
  summary()
}
}
