% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_labels.R
\name{set_labels}
\alias{set_labels}
\title{Add \code{label} attribute to columns in a dataframe}
\usage{
set_labels(.data, .cols, .labels = NULL)
}
\arguments{
\item{.data}{A dataframe.}

\item{.cols}{A character vector. Colomns to add labels.}

\item{.labels}{A character vector. Labels for the columns. If NULL, each element of .colnames must be named with its label.}
}
\value{
Dataframe with labelled columns.
}
\description{
Add \code{label} attribute to columns in a dataframe
}
\examples{
\dontrun{
dat <- data.frame(
  age = c(23, 45),
  gender = c("M", "F")
)

# Labels in a separate vector
var <- c("age", "gender")
lbl <- c("Age of participant", "Gender of participant")
set_labels(dat, var, lbl)

# Labels in the names of `.cols` vector
names(var) <- lbl
set_labels(dat, var)
}
}
