% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Create project folders}
\usage{
create_folders(project_path = ".")
}
\arguments{
\item{project_path}{Project folder path.}
}
\value{

}
\description{
Create project folders with a minimal structure:
code - R script files (.R)
data_raw - Raw data from external source
data_derived - Data files generated in the analysis process
figs - Figures and plots
docs - Report files (.Rmd and their derivatives or .docx)
}
\details{
In a typical new workflow setup, create a RStudio project folder first. Then,
run this function in R console using the default parameter.
}
\examples{
temp_folder <- tempdir()
create_folders(temp_folder)
list.dirs(temp_folder)
}
