% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data.R
\name{describe_data}
\alias{describe_data}
\title{Create a tableby object using {arsenal} package with more intuitive default parameters. The tableby object should be passed to summary() to create a descriptive table.}
\usage{
describe_data(dat, by = "", p_value = FALSE, total = FALSE, ...)
}
\arguments{
\item{dat}{A dataframe}

\item{by}{A string with length 1. Name of grouping variable by which data are described. If "" (default), overall total will be shown.}

\item{p_value}{A logical vector with length 1. Toggle to show p value (default is FALSE).}

\item{total}{A logical vector with length 1. Toggle to show overall total (default is FALSE) when grouping variable is defined.}

\item{...}{Arguments passed to tableby(). ?tableby for more details.}
}
\value{

}
\description{
Create a tableby object using {arsenal} package with more intuitive default parameters. The tableby object should be passed to summary() to create a descriptive table.
}
\examples{
\dontrun{
library(dplyr)
data(mockstudy)

mockstudy \%>\%
  select(arm, sex, age) \%>\%
  describe_data(by = "arm") \%>\%
  summary(text = TRUE)
}
}
